feat: add favicon
