+++ src/scss/_question-block.scss
@@ -11,6 +11,7 @@
   &__image {
     width: 230px;
     height: 180px;
+    border-radius: 10px;
   }
 
   &__description {
